[ADD] fmt tests and change files from varint to zigzag
